some minor bug fixes

+++ trunk/csharp/core/media/data/audio/PCMFormatInfo.cs
@@ -159,12 +159,13 @@ namespace urakawa.media.data.audio
         /// </summary>
         /// <param name="dataLen">The length</param>
         /// <returns>The duration</returns>
-        public TimeDelta GetDuration(uint dataLen)
+        public TimeDelta GetDuration(double dataLen)
         {
             if (ByteRate == 0)
             {
                 throw new exception.InvalidDataFormatException("The PCM data has byte rate 0");
             }
+
             double blockCount = dataLen/BlockAlign;
             return new TimeDelta(TimeSpan.FromTicks((long) (Math.Round(GetTicksPerBlock()*blockCount))));
         }
